move Keycloak Classes
