+++ src/main/resources/templates/Orga/bewerbungenVerwalten.html
@@ -1,5 +1,5 @@
 <!DOCTYPE html>
-<html lang="en">
+<html lang="en" xmlns:th="http://www.w3.org/1999/xhtml">
 <head>
   <meta charset="UTF-8">
   <title>bewerbungenVerwalten</title>
@@ -8,6 +8,8 @@
 
 <h1>Autorisierter Mitarbeiter verwaltet die Bewerbungen</h1>
 <h2>Zugang muss weiter eingeschränkt werden</h2>
+
+<p th:text="${dozentmail}">Nicht angemeldet</p>
 
 </body>
 </html>
